Move cookie logic to background service worker
